django/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0